CR fixes: 1. Do not copy qcow2 to tmp location - use already stored one 2. Fix transport_test.go test

+++ pkg/importer/dataStream.go
@@ -65,13 +65,15 @@ var qemuOperations = image.NewQEMUOperations()
 // DataStream implements the ReadCloser interface
 type DataStream struct {
 	*DataStreamOptions
-	url     *url.URL
-	Readers []reader
-	buf     []byte // holds file headers
-	qemu    bool
-	Size    int64
-	ctx     context.Context
-	cancel  context.CancelFunc
+	url         *url.URL
+	Readers     []reader
+	buf         []byte // holds file headers
+	qemu        bool
+	archived    bool
+	tmpDataPath string
+	Size        int64
+	ctx         context.Context
+	cancel      context.CancelFunc
 }
 
 type reader struct {
@@ -122,6 +124,30 @@ var rdrTypM = map[string]int{
 	"stream": rdrStream,
 }
 
+//Data stream helper methods
+func (d *DataStream) isCopyToTmpRequired() bool {
+	if d.isQemu() && d.isStored() && !d.isArchived() {
+		return false
+	}
+	return true
+}
+
+func (d *DataStream) getTmpDataPath() string {
+	return d.tmpDataPath
+}
+
+func (d *DataStream) isStored() bool {
+	return !(len(d.tmpDataPath) == 0)
+}
+
+func (d *DataStream) isQemu() bool {
+	return d.qemu
+}
+
+func (d *DataStream) isArchived() bool {
+	return d.archived
+}
+
 // NewDataStream returns a DataStream object after validating the endpoint and constructing the reader/closer chain.
 // Note: the caller must close the `Readers` in reverse order. See Close().
 func NewDataStream(dso *DataStreamOptions) (*DataStream, error) {
@@ -162,6 +188,9 @@ func newDataStream(dso *DataStreamOptions, stream io.ReadCloser) (*DataStream, e
 		buf:               make([]byte, image.MaxExpectedHdrSize),
 		ctx:               ctx,
 		cancel:            cancel,
+		tmpDataPath:       "",
+		qemu:              false,
+		archived:          false,
 	}
 
 	// establish readers for endpoint's formats and do initial calc of size of raw endpt
@@ -199,9 +228,10 @@ func (d *DataStream) Close() error {
 
 // Based on the endpoint scheme, append the scheme-specific reader to the receiver's
 // reader stack.
-func (d *DataStream) dataStreamSelector() (err error) {
+func (d *DataStream) dataStreamSelector() error {
 	var r io.Reader
 	scheme := d.url.Scheme
+	var err error
 	switch scheme {
 	case "s3":
 		r, err = d.s3()
@@ -210,8 +240,10 @@ func (d *DataStream) dataStreamSelector() (err error) {
 	case "docker", "oci":
 		r, err = d.registry()
 	default:
+		glog.V(1).Infoln("Error in dataStream Selector - invalid url scheme")
 		return errors.Errorf("invalid url scheme: %q", scheme)
 	}
+
 	if err == nil && r != nil {
 		d.appendReader(rdrTypM[scheme], r)
 	}
@@ -278,6 +310,18 @@ type registryData struct {
 	file    *os.File
 }
 
+func (r *registryData) Close() error {
+	glog.V(1).Infof("registryData - deleting all the data")
+	if _regData.file != nil {
+		_regData.file.Close()
+	}
+	err := os.RemoveAll(_regData.dataDir)
+	if err != nil {
+		glog.Errorf(" Failed removing directory ", err.Error())
+	}
+	return nil
+}
+
 //_regData - global variable that stores temporary data of registy flow
 // to be deleted when teh flow ends
 var _regData = registryData{"", nil}
@@ -287,9 +331,7 @@ type registryDataCloser struct {
 }
 
 func (registryDataCloser) Close() error {
-	_regData.file.Close()
-	os.RemoveAll(_regData.dataDir)
-	return nil
+	return _regData.Close()
 }
 
 // RegistryDataCloser returns a ReadCloser with a Close method wrapping delettion
@@ -319,6 +361,7 @@ func (d *DataStream) registry() (io.ReadCloser, error) {
 	if _, err := os.Stat(_regData.dataDir); os.IsNotExist(err) {
 		err := os.Mkdir(_regData.dataDir, os.ModeDir|os.ModePerm)
 		if err != nil {
+			glog.Errorf("Failed to create temporary directory")
 			return nil, errors.Wrapf(err, fmt.Sprintf("Failed to create tempdirectory %s", _regData.dataDir))
 		}
 	}
@@ -327,25 +370,34 @@ func (d *DataStream) registry() (io.ReadCloser, error) {
 	glog.V(1).Infof("using skopeo to copy from registry")
 	err := image.CopyDirFromRegistryImage(d.Endpoint, _regData.dataDir, ContainerDiskImageDir, d.AccessKey, d.SecKey)
 	if err != nil {
-		os.RemoveAll(_regData.dataDir)
+		_regData.Close()
+		glog.Errorf("Failed to read data from registry")
 		return nil, errors.Wrapf(err, fmt.Sprintf("Failed ro read from registry"))
 	}
 
 	//Search for file in /disk directory - if not found - failure
 	imageFile, err := getImageFileName(imageDir)
 	if err != nil {
-		os.RemoveAll(_regData.dataDir)
+		_regData.Close()
+		glog.Errorf("Error getting Image file from imageDirectory")
 		return nil, errors.Wrapf(err, fmt.Sprintf("Cannot locate image file"))
 	}
 
+	if len(strings.TrimSpace(imageFile)) == 0 {
+		_regData.Close()
+		glog.Errorf("Image file is not found in image directory - filename is empty")
+		return nil, errors.Errorf(fmt.Sprintf("Cannot locate image file"))
+	}
 	// 3. If found - Create a reader that will read this file and attach it to the dataStream
 	_regData.file, err = os.Open(filepath.Join(imageDir, imageFile))
 	if err != nil {
-		os.RemoveAll(_regData.dataDir)
+		_regData.Close()
+		glog.Errorf("Failed to open image file")
 		return nil, errors.Wrapf(err, fmt.Sprintf("Fail to create data stream from image file"))
 	}
 
-	glog.V(1).Infof("VM disk image filename is %s", imageFile)
+	d.tmpDataPath = filepath.Join(imageDir, imageFile)
+	glog.V(1).Infof("Sucecssfully found file. VM disk image filename is %s", imageFile)
 
 	return RegistryDataCloser(bufio.NewReader(_regData.file)), nil
 }
@@ -353,23 +405,32 @@ func (d *DataStream) registry() (io.ReadCloser, error) {
 func getImageFileName(dir string) (string, error) {
 
 	if _, err := os.Stat(dir); os.IsNotExist(err) {
-		return "", errors.Wrapf(err, fmt.Sprintf("image directory does not exist "))
+		glog.Errorf("image directory does not exist")
+		return "", errors.Errorf(fmt.Sprintf("image directory does not exist "))
 	}
 
 	entries, err := ioutil.ReadDir(dir)
-	if len(entries) == 0 || err != nil {
+	if err != nil {
+		glog.Errorf("Error reading directory")
 		return "", errors.Wrapf(err, fmt.Sprintf("image file does not exist in image directory"))
+	}
+
+	if len(entries) == 0 {
+		glog.Errorf("image file does not exist in image directory - directory is empty ")
+		return "", errors.Errorf(fmt.Sprintf("image file does not exist in image directory - directory is empty"))
 	}
 
 	fileinfo := entries[len(entries)-1]
 	if fileinfo.IsDir() {
-		return "", errors.Wrapf(err, fmt.Sprintf("image file does not exist in image directory"))
+		glog.Errorf("image file does not exist in image directory contains another directory ")
+		return "", errors.Errorf(fmt.Sprintf("image file does not exist in image directory"))
 	}
 
 	filename := fileinfo.Name()
 
 	if len(strings.TrimSpace(filename)) == 0 {
-		return "", errors.Wrapf(err, fmt.Sprintf("image file does not exist in image directory"))
+		glog.Errorf("image file does not exist in image directory - file has no name ")
+		return "", errors.Errorf(fmt.Sprintf("image file does not exist in image directory"))
 	}
 
 	glog.V(1).Infof("VM disk image filename is %s", filename)
@@ -402,21 +463,19 @@ func (d *DataStream) pollProgress(reader *util.CountingReader, idleTime, pollInt
 // CopyData copies the source endpoint (vm image) to the provided destination path.
 func CopyData(dso *DataStreamOptions) error {
 	glog.V(1).Infof("copying %q to %q...\n", dso.Endpoint, dso.Dest)
-	switch dso.Source {
-	default:
-		ds, err := NewDataStream(dso)
-		if err != nil {
-			return errors.Wrap(err, "unable to create data stream")
-		}
-		defer ds.Close()
-		if dso.ContentType == string(cdiv1.DataVolumeArchive) {
-			if err := util.UnArchiveTar(ds.topReader(), dso.Dest); err != nil {
-				return errors.Wrap(err, "unable to untar files from endpoint")
-			}
-			return nil
-		}
-		return ds.copy(dso.Dest)
+	ds, err := NewDataStream(dso)
+	if err != nil {
+		return errors.Wrap(err, "unable to create data stream")
 	}
+	defer ds.Close()
+	if dso.ContentType == string(cdiv1.DataVolumeArchive) {
+		if err := util.UnArchiveTar(ds.topReader(), dso.Dest); err != nil {
+			return errors.Wrap(err, "unable to untar files from endpoint")
+		}
+		return nil
+	}
+	return ds.copy(dso.Dest)
+
 }
 
 // SaveStream reads from a stream and saves data to dest
@@ -500,6 +559,7 @@ func (d *DataStream) constructReaders(stream io.ReadCloser) error {
 		// create the scheme-specific source reader and append it to dataStream readers stack
 		err := d.dataStreamSelector()
 		if err != nil {
+			glog.Errorf("failed to construct dataStream from endpoint")
 			return errors.WithMessage(err, "could not get data reader")
 		}
 	} else {
@@ -580,13 +640,22 @@ func (d *DataStream) fileFormatSelector(hdr *image.Header) (err error) {
 	switch fFmt {
 	case "gz":
 		r, d.Size, err = d.gzReader()
+		if err != nil {
+			d.archived = true
+		}
 	case "qcow2":
 		r, d.Size, err = d.qcow2NopReader(hdr)
 		d.qemu = true
 	case "tar":
 		r, d.Size, err = d.tarReader()
+		if err != nil {
+			d.archived = true
+		}
 	case "xz":
 		r, d.Size, err = d.xzReader()
+		if err != nil {
+			d.archived = true
+		}
 	default:
 		return errors.Errorf("mismatch between supported file formats and this header type: %q", fFmt)
 	}
@@ -840,21 +909,26 @@ func (d *DataStream) copy(dest string) error {
 		dest = filepath.Clean(dest)
 		glog.V(2).Infof("copying image file to %q", dest)
 		tmpDest := dest
-		if d.qemu {
-			// copy to tmp; qemu conversion will write to passed-in destination
-			tmpDest = randTmpName(dest)
-			glog.V(3).Infof("Copy: temp file for qcow2 conversion: %q", tmpDest)
-			defer func(f string) {
-				os.Remove(f)
-			}(tmpDest)
+		if d.isCopyToTmpRequired() {
+			if d.qemu {
+				// copy to tmp; qemu conversion will write to passed-in destination
+				tmpDest = randTmpName(dest)
+				glog.V(3).Infof("Copy: temp file for qcow2 conversion: %q", tmpDest)
+				defer func(f string) {
+					os.Remove(f)
+				}(tmpDest)
+			}
+
+			err := StreamDataToFile(d.topReader(), tmpDest)
+			if err != nil {
+				return errors.WithMessage(err, fmt.Sprintf("unable to stream data to file %q", dest))
+			}
+		} else {
+			tmpDest = d.getTmpDataPath()
 		}
 		// The actual copy
-		err := StreamDataToFile(d.topReader(), tmpDest)
-		if err != nil {
-			return errors.WithMessage(err, fmt.Sprintf("unable to stream data to file %q", dest))
-		}
 		if d.qemu {
-			err = d.convertQcow2ToRaw(tmpDest, dest)
+			err := d.convertQcow2ToRaw(tmpDest, dest)
 			if err != nil {
 				return err
 			}
